Add files via upload
This was a group Project in CoderHouse, it was qualified with a 10/10 and reached the top 10% of the class! 
We decided to study how governments spend their money, and how that affects the economic development of each country.
By doing that, we were able to understand in which sectors the public investments have the better results in the long term.